working on latlon detection
Need to think more about + and - for elevation, and how to handle in latlon inference.  (Fits are a
lot better when only either lat or lon inferred).  And what, exactly, do the times mean -- are they
the times when the sun arc hits the horizon, or the sun centre, or the opposite arc?